I took a profiler to this package and got a bunch of fun performance wins!

The net improvement from commits in this PR (as recorded on my early 2013 2.7 GHz Core i7 MacBook Pro) is as follows:
```
benchmark             old ns/op     new ns/op     delta
BenchmarkHash1024     11766         4608          -60.84%

benchmark             old MB/s     new MB/s     speedup
BenchmarkHash1024     87.03        222.21       2.55x

benchmark             old allocs     new allocs     delta
BenchmarkHash1024     66             0              -100.00%

benchmark             old bytes     new bytes     delta
BenchmarkHash1024     2064          0             -100.00%
```

The bulk of my effort was put into reducing allocations, so I'm sure I left some cycles on the table if you (or anyone else) is curious to hunt for them.